ENH: Remove unused default process count parameter
Remove unused default process count parameter from the process count
validation method.

Take advantage of the commit to:
- Change the "less than or equal to" 0 condition on the value to be
returned to a "strictly less than": if a value of 0 is provided in the
arguments, then the returned value corresponds to the available core
count, and thus the value will never be less than 1.
- Improve the method documentation: state clearly the range of valid
values and document the method behavior if the process count is not
contained in the arguments.
- Fix the returned value docstring.
- Fix the logged process count in script.